vkd3d-shader/hlsl: Introduce specialized helpers for arithmetic operations.
In HLSL the rules used to determine the type of a composite expression
change depending on the operator. This commit changes add_expr so that
it accepts both the types to which the operands have to be casted
and the type that the final expression must have (which is not
necessarily the cast type).

For example, when comparing a half with an int, both of them will
have to be casted to float, though the type of the expression as a
whole will be bool.

Similarly, it is not always the case that all operands have to be
casted to the same type, therefore the interface leaves room for
specifying a different cast type for each operand. For example,
the first operand of the ternary operator must be casted to bool,
while the other two must be casted to their common type, which is
then the output type of the whole expression.

Signed-off-by: Giovanni Mascellani <gmascellani@codeweavers.com>