vkd3d-shader/preproc: Always emit \n for T_NEWLINE.
This has the effects of removing \r from the output, and of no longer emitting
spaces at the beginning of a line.

The motivation for the patch is the second; our HLSL lexer is not very liberal
about what it considers a #pragma. Although our HLSL lexer is arguably overly
restrictive, native d3dcompiler doesn't emit spaces at the beginning of a line,
and therefore we might as well fix it here.